refactor: move test files, remove old snapshot
