+++ src/overpass/__tests__/fetch-data.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { getFetchUrl } from "./fetch-data.js";
+import { getFetchUrl } from "../fetch-data.js";
 
 describe("getFetchUrl function", () => {
   it("should URL read overpassQL query correctly", async () => {
