--- src/overpass/fetch-data.test.js.snapshot
@@ -1,3 +1,0 @@
-exports[`getFetchUrl function > should URL read overpassQL query correctly 1`] = `
-"https://overpass-api.de/api/interpreter?data=%0A%5Bout%3Axml%5D%5Btimeout%3A25%5D%3B%0A%2F%2F+fetch+area+%E2%80%9Cbergen%E2%80%9D+to+search+in%0Aarea%28id%3A3600404159%29-%3E.searchArea%3B%0A%0A%2F%2F+find+second+hand+shops%0A%28%0A++node%5B%22shop%22%3D%22second_hand%22%5D%28area.searchArea%29%3B%0A++node%5B%22second_hand%22%3D%22yes%22%5D%28area.searchArea%29%3B%0A++node%5B%22second_hand%22%3D%22only%22%5D%28area.searchArea%29%3B%0A++node%5B%22amenity%22%3D%22give_box%22%5D%28area.searchArea%29%3B%0A%29-%3E.secondHand%3B%0A%0A%2F%2F+send+final+output+to+default+result+set+%0A%28%0A++node.secondHand%3B%0A%29%3B%0A%0Aout+geom%3B"
-`;
